# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: APEX.BUILD API
  description: |
    Next-generation cloud development platform with multi-AI integration.

    APEX.BUILD provides:
    - Multi-AI integration (Claude, GPT-4, Gemini, Grok, Ollama)
    - Real-time code collaboration
    - Cloud code execution with Docker container sandboxing
    - One-click deployment (Vercel, Netlify, Render, native hosting)
    - Live preview with hot reload
    - BYOK (Bring Your Own Key) for AI providers

    ## Authentication

    All protected endpoints require a Bearer JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    Obtain tokens via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting

    - Authentication endpoints: 10 requests/minute
    - AI endpoints: Based on subscription tier
    - General API: 1000 requests/hour

    ## WebSocket Endpoints

    Real-time features use WebSocket connections:
    - `/ws/build/:buildId` - Build progress updates
    - `/ws/terminal/:sessionId` - Interactive terminal
    - `/ws/collab` - Real-time collaboration
    - `/ws/debug/:sessionId` - Debug session
    - `/ws/deploy/:deploymentId` - Deployment logs
    - `/mcp/ws` - MCP protocol

  version: 1.0.0
  contact:
    name: APEX.BUILD Support
    email: support@apex.build
    url: https://apex.build
  license:
    name: Proprietary
    url: https://apex.build/terms
servers:
  - url: https://api.apex.build/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /deploy:
    post:
      tags: [Deploy]
      summary: Start deployment
      description: Deploy project to cloud provider (Vercel, Netlify, Render, Railway, Cloudflare Pages)
      operationId: startDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartDeploymentRequest'
      responses:
        '202':
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
  /deploy/{id}:
    get:
      tags: [Deploy]
      summary: Get deployment details
      operationId: getDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
    delete:
      tags: [Deploy]
      summary: Cancel deployment
      operationId: cancelDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /deploy/{id}/status:
    get:
      tags: [Deploy]
      summary: Get deployment status
      operationId: getDeploymentStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                  url:
                    type: string
  /deploy/{id}/logs:
    get:
      tags: [Deploy]
      summary: Get deployment logs
      operationId: getDeploymentLogs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Deployment logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
  /deploy/{id}/redeploy:
    post:
      tags: [Deploy]
      summary: Redeploy
      operationId: redeploy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Redeployment started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
...
  /deploy/projects/{projectId}/history:
    get:
      tags: [Deploy]
      summary: Get deployment history
      operationId: getProjectDeployments
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Deployment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  total:
                    type: integer

  # ==================== HOSTING ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Project ID
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    StartDeploymentRequest:
      type: object
      required: [project_id, provider]
      properties:
        project_id:
          type: integer
        provider:
          type: string
          enum: [vercel, netlify, render, railway, cloudflare_pages]
        environment:
          type: string
          default: production
        branch:
          type: string
          default: main
        env_vars:
          type: object
          additionalProperties:
            type: string
        build_command:
          type: string
        output_dir:
          type: string
        install_cmd:
          type: string
        start_command:
          type: string
        framework:
          type: string
        node_version:
          type: string
        root_directory:
          type: string
        database:
          type: object
          properties:
            provider:
              type: string
              enum: [neon]
            project_name:
              type: string
            branch_name:
              type: string
            database_name:
              type: string
            role_name:
              type: string
            region_id:
              type: string
            org_id:
              type: string
            pg_version:
              type: integer
              enum: [14, 15, 16, 17]
            pooled:
              type: boolean
    Deployment:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: integer
        user_id:
          type: integer
        provider:
          type: string
        status:
          type: string
          enum: [pending, preparing, building, deploying, live, failed, cancelled]
        url:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        deployed_at:
          type: string
          format: date-time
    DeploymentResponse:
      type: object
      properties:
        success:
          type: boolean
        deployment:
          $ref: '#/components/schemas/Deployment'
        message:
          type: string
        websocket_url:
          type: string
    LogEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        source:
          type: string
        message:
          type: string

    # ==================== Hosting ====================
...